Add a way to keep the test outstore around for inspection

+++ jenkins/jenkins.sh
@@ -19,6 +19,8 @@
 LOCAL_BUILD=0
 # Should we re-use and keep around the same virtualenv?
 REUSE_VENV=0
+# Should we keep our test output around after uploading the new baseline?
+KEEP_OUTPUT=0
 # What toil-vg should we install?
 TOIL_VG_PACKAGE="git+https://github.com/adamnovak/toil-vg.git@563fd24048b6dc1b1b09e86443b9a9aea9acc996"
 # What tests should we run?
@@ -33,18 +35,22 @@ usage() {
     printf "\t-l\t\tBuild vg locally (instead of in Docker) and don't use Docker at all.\n"
     printf "\t\t\tNon-Python dependencies must be installed.\n"
     printf "\t-r\t\tRe-use a single virtualenv. \n"
+    printf "\t-k\t\tKeep output. \n"
     printf "\t-p PACKAGE\tUse the given Python package specifier to install toil-vg.\n"
     printf "\t-t TESTSPEC\tUse the given PyTest test specifier to select tests to run.\n"
     exit 1
 }
 
-while getopts "lrp:t:" o; do
+while getopts "lrkp:t:" o; do
     case "${o}" in
         l)
             LOCAL_BUILD=1
             ;;
         r)
             REUSE_VENV=1
+            ;;
+        k)
+            KEEP_OUTPUT=1
             ;;
         p)
             TOIL_VG_PACKAGE="${OPTARG}"
@@ -213,12 +219,13 @@ if [ ! "${REUSE_VENV}" == "1" ]; then
     rm -rf awscli s3am
 fi
 
-if [ "${LOCAL_BUILD}" == "0" ] || [ "${PYRET}" == 0 ]; then
-    # On anything other than a failed local run, clean up.
+if ([ "${LOCAL_BUILD}" == "0" ] || [ "${PYRET}" == 0 ]) && [ ! "${KEEP_OUTPUT}" == "1" ]; then
+    # On anything other than a failed local run, and if we haven't been told not to, clean up the test output.
     rm -rf vgci-work
-    if [ ! "${REUSE_VENV}" == "1" ]; then
-        rm -rf .env
-    fi
+fi
+if [ ! "${REUSE_VENV}" == "1" ]; then
+    # If we aren't re-using the virtualenv, clean it up
+    rm -rf .env
 fi
 
 if [ -d "/mnt/ephemeral" ]
